pyproject.toml
[project]
name = "send-money-agent"
version = "0.1.0"
description = "Money transfer agent built with Google ADK"
requires-python = ">=3.11"
dependencies = [
    "google-adk[extensions]>=1.0.0",
    "python-dotenv",
    "loguru",
    "fastapi",
    "uvicorn[standard]",
    "sse-starlette",
    "aiosqlite",
    "httpx",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
